Remove unneeded scope at packed_xml.js, move unpacked XML processing logic into transport.js
