+++ lib/util/transport.js
@@ -1,4 +1,5 @@
-var packed_xml = require('./packed_xml.js');
+var packed_xml = require('./packed_xml.js'),
+    b64_decode = require('./base64.js').b64_decode;
 
 var write_buf = function(buf, data) {
     if(!data.length) {
@@ -150,6 +151,42 @@ var qlb_split = function (buffer, print) {
     return ret;
 };
 
+var decode_packed_xml_contents = function(xmlBuf) {
+    var el, partName;
+    var res = {};
+    // Regexp to find a pair of HTML tags.
+    // [\s\S] inside a tag matches all white-space characters + all non-whitespace
+    // characters, which gives us all characters including newline.
+    // The standard dot (.) does not match newline, and there is no 's' flag in JS
+    var regexp = new RegExp('<([\\w][\\w\\d]*)([^>]*)>([\\s\\S]*?)</\\1>', 'g');
+    var rootElement, childElement;
+
+    if (xmlBuf.slice(0, 5).toString() !== "<?xml") {
+        throw Error("This is not XML");
+    }
+
+    // first find root tag
+    rootElement = regexp.exec(xmlBuf.toString());
+    if (rootElement === null) {
+        throw Error("Invalid XML data - cannot find root element");
+    }
+    if (rootElement.length < 4) {
+        throw Error("Invalid XML data - cannot process root element");
+    }
+
+    // now go through it's elements
+    regexp.lastIndex = 0;
+    while ((childElement = regexp.exec(rootElement[3])) !== null) {
+        if (childElement.length < 4) {
+            continue;
+        }
+        partName = childElement[1] + childElement[2];
+        res[partName] = b64_decode(childElement[3]);
+    }
+
+    return res;
+}
+
 var transport_decode = function (buffer) {
     var ret = {docs: []};
     var off = 0;
@@ -162,8 +199,15 @@ var transport_decode = function (buffer) {
     while (off < buffer.length) {
 
         if (buffer.slice(0, 10).toString() === "PACKED_XML") {
-            unpacker = packed_xml.Unpacker();
-            ct = unpacker.unpack(buffer);
+            try {
+                ct = packed_xml.unpack(buffer);
+                ct = decode_packed_xml_contents(ct);
+            } catch (e) {
+                // let the user know that something wrong,
+                // as exception will be eaten later...
+                console.error(e.message);
+                throw e;
+            }
             ret.header = {};
             for (var doc in ct) {
                 if (doc === "DOCUMENT") {
